Social: Add an endpoint to generate the token for the SIG preview (#29596)
* [not verified] Social: Add an endpoint to generate the token for the SIG preview

So we can display a preview of the generated image, we need to call the
WPCOM endpoint to generate the token to pass to the generator service.
This change adds an endpoint to the Jetpack site, that calls the WPCOM
endpoint and returns the resulting token.

* [not verified] Refactor to move the endpoint to a separate class.

* Added some tests.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4510091717